refactor(calendar): インタラクション状態機械と統合 hook を責務分割する
machine.ts（644行）を reducer 本体 + machine-constants / grid-geometry /
pointer-move / pointer-up に、useInteraction.ts（604行）を hook 本体 +
interaction-runtime（型） / interaction-effects（副作用実行） /
useInteractionListeners（DOM リスナー・カーソル）に分割。
ロジックは書き換えずに移動のみ。machine.ts の既存 export は re-export で維持。
machine.test.ts（1306行）も 6 ファイル + 共有ヘルパーへ追随分割。

Refs #1527
