+++ apps/product/src/features/calendar/domain/interaction/__tests__/machine-precision.test.ts
@@ -1,0 +1,351 @@
+/**
+ * Interaction State Machine tests — snapToGrid と精度・日端境界の回帰
+ */
+
+import { describe, expect, it } from 'vitest';
+
+import { IDLE, snapToGrid } from '../machine';
+import type { InteractionState, TimeblockRect } from '../types';
+import { createCtx, dispatch, origin } from './machine-test-helpers';
+
+// ========================================
+// snapToGrid
+// ========================================
+
+describe('snapToGrid', () => {
+  it('snaps to 15-min intervals', () => {
+    // 9:00 = 540px at 60px/h
+    expect(snapToGrid(540, 60)).toEqual({ snappedTop: 540, hour: 9, minute: 0 });
+  });
+
+  it('snaps 9:07 → 9:00', () => {
+    // 9:07 = 547px
+    expect(snapToGrid(547, 60)).toEqual({ snappedTop: 540, hour: 9, minute: 0 });
+  });
+
+  it('snaps 9:08 → 9:15', () => {
+    // 9:08 = 548px
+    expect(snapToGrid(548, 60)).toEqual({ snappedTop: 555, hour: 9, minute: 15 });
+  });
+
+  it('snaps 9:52 → 9:45', () => {
+    // 9:52 = 592px
+    expect(snapToGrid(592, 60, 15)).toEqual({ snappedTop: 585, hour: 9, minute: 45 });
+  });
+
+  it('snaps 9:53 → 10:00', () => {
+    // 9:53 = 593px
+    expect(snapToGrid(593, 60, 15)).toEqual({ snappedTop: 600, hour: 10, minute: 0 });
+  });
+
+  it('clamps to 23:00 max', () => {
+    expect(snapToGrid(9999, 60)).toEqual({ snappedTop: 1380, hour: 23, minute: 0 });
+  });
+
+  it('clamps to 0:00 min', () => {
+    expect(snapToGrid(-10, 60)).toEqual({ snappedTop: 0, hour: 0, minute: 0 });
+  });
+
+  it('supports custom interval (30 min)', () => {
+    // 9:20 → 9:30
+    expect(snapToGrid(560, 60, 30)).toEqual({ snappedTop: 570, hour: 9, minute: 30 });
+  });
+});
+describe('precision regression: drag/resize snaps off-grid times to the absolute grid', () => {
+  // 10:07 entry を表す originalPosition: top=607（hourHeight=60 で 10*60+7=607）
+  const off607: TimeblockRect = { top: 607, left: 0, width: 200, height: 60 };
+
+  it('LONGPRESS_FIRED on 10:07 entry → preview start snaps to 10:00', () => {
+    const longpressState: InteractionState = {
+      mode: 'longpress-pending',
+      timeblockId: 'a',
+      startPoint: origin,
+      originalPosition: off607,
+      dateIndex: 0,
+    };
+    const { state } = dispatch(longpressState, { type: 'LONGPRESS_FIRED' });
+
+    expect(state.mode).toBe('dragging');
+    if (state.mode === 'dragging') {
+      expect(state.snappedTop).toBe(600);
+      expect(state.previewTime.start.getHours()).toBe(10);
+      expect(state.previewTime.start.getMinutes()).toBe(0);
+    }
+  });
+
+  it('drag 10:07 → +30px → 10:30', () => {
+    const draggingState: InteractionState = {
+      mode: 'dragging',
+      timeblockId: 'a',
+      startPoint: origin,
+      currentPoint: origin,
+      originalPosition: off607,
+      dateIndex: 0,
+      targetDateIndex: 0,
+      snappedTop: 607,
+      previewTime: {
+        start: new Date('2026-01-15T10:07:00'),
+        end: new Date('2026-01-15T11:07:00'),
+      },
+      isOverlapping: false,
+    };
+    const movedPoint = { clientX: origin.clientX, clientY: origin.clientY + 30 };
+    const { state } = dispatch(draggingState, { type: 'POINTER_MOVE', point: movedPoint });
+
+    if (state.mode === 'dragging') {
+      expect(state.snappedTop).toBe(630);
+      expect(state.previewTime.start.getHours()).toBe(10);
+      expect(state.previewTime.start.getMinutes()).toBe(30);
+    }
+  });
+
+  it('drag 10:07 → +7px → 10:15', () => {
+    const draggingState: InteractionState = {
+      mode: 'dragging',
+      timeblockId: 'a',
+      startPoint: origin,
+      currentPoint: origin,
+      originalPosition: off607,
+      dateIndex: 0,
+      targetDateIndex: 0,
+      snappedTop: 607,
+      previewTime: {
+        start: new Date('2026-01-15T10:07:00'),
+        end: new Date('2026-01-15T11:07:00'),
+      },
+      isOverlapping: false,
+    };
+    const movedPoint = { clientX: origin.clientX, clientY: origin.clientY + 7 };
+    const { state } = dispatch(draggingState, { type: 'POINTER_MOVE', point: movedPoint });
+
+    if (state.mode === 'dragging') {
+      expect(state.snappedTop).toBe(615);
+      expect(state.previewTime.start.getMinutes()).toBe(15);
+    }
+  });
+
+  it('5分snap設定では15:13 entryをdragすると15:15へ揃える', () => {
+    const off913: TimeblockRect = { top: 913, left: 0, width: 200, height: 60 };
+    const draggingState: InteractionState = {
+      mode: 'dragging',
+      timeblockId: 'a',
+      startPoint: origin,
+      currentPoint: origin,
+      originalPosition: off913,
+      dateIndex: 0,
+      targetDateIndex: 0,
+      snappedTop: 913,
+      previewTime: {
+        start: new Date('2026-01-15T15:13:00'),
+        end: new Date('2026-01-15T16:13:00'),
+      },
+      isOverlapping: false,
+    };
+    const { state } = dispatch(
+      draggingState,
+      { type: 'POINTER_MOVE', point: origin },
+      createCtx({ snapIntervalMinutes: 5 }),
+    );
+
+    if (state.mode === 'dragging') {
+      expect(state.snappedTop).toBe(915);
+      expect(state.previewTime.start.getHours()).toBe(15);
+      expect(state.previewTime.start.getMinutes()).toBe(15);
+      expect(state.previewTime.end.getHours()).toBe(16);
+      expect(state.previewTime.end.getMinutes()).toBe(15);
+    }
+  });
+
+  it('5分snap設定ではdrag時にoff-grid durationも丸めて終了時刻をグリッドへ揃える', () => {
+    const off913: TimeblockRect = { top: 913, left: 0, width: 200, height: 47 };
+    const draggingState: InteractionState = {
+      mode: 'dragging',
+      timeblockId: 'a',
+      startPoint: origin,
+      currentPoint: origin,
+      originalPosition: off913,
+      dateIndex: 0,
+      targetDateIndex: 0,
+      snappedTop: 913,
+      previewTime: {
+        start: new Date('2026-01-15T15:13:00'),
+        end: new Date('2026-01-15T16:00:00'),
+      },
+      isOverlapping: false,
+    };
+    const { state } = dispatch(
+      draggingState,
+      { type: 'POINTER_MOVE', point: origin },
+      createCtx({
+        snapIntervalMinutes: 5,
+        getTimeblockDurationMs: () => 47 * 60 * 1000,
+      }),
+    );
+
+    if (state.mode === 'dragging') {
+      const durationMs = state.previewTime.end.getTime() - state.previewTime.start.getTime();
+      expect(state.previewTime.start.getHours()).toBe(15);
+      expect(state.previewTime.start.getMinutes()).toBe(15);
+      expect(state.previewTime.end.getHours()).toBe(16);
+      expect(state.previewTime.end.getMinutes()).toBe(0);
+      expect(durationMs).toBe(45 * 60 * 1000);
+    }
+  });
+
+  it('drag時は終了境界もsnapし、48分entryを不要に50分へ伸ばさない', () => {
+    const off913: TimeblockRect = { top: 913, left: 0, width: 200, height: 48 };
+    const draggingState: InteractionState = {
+      mode: 'dragging',
+      timeblockId: 'a',
+      startPoint: origin,
+      currentPoint: origin,
+      originalPosition: off913,
+      dateIndex: 0,
+      targetDateIndex: 0,
+      snappedTop: 913,
+      previewTime: {
+        start: new Date('2026-01-15T15:13:00'),
+        end: new Date('2026-01-15T16:01:00'),
+      },
+      isOverlapping: false,
+    };
+    const { state } = dispatch(
+      draggingState,
+      { type: 'POINTER_MOVE', point: origin },
+      createCtx({
+        snapIntervalMinutes: 5,
+        getTimeblockDurationMs: () => 48 * 60 * 1000,
+      }),
+    );
+
+    if (state.mode === 'dragging') {
+      const durationMs = state.previewTime.end.getTime() - state.previewTime.start.getTime();
+      expect(state.previewTime.start.getHours()).toBe(15);
+      expect(state.previewTime.start.getMinutes()).toBe(15);
+      expect(state.previewTime.end.getHours()).toBe(16);
+      expect(state.previewTime.end.getMinutes()).toBe(0);
+      expect(durationMs).toBe(45 * 60 * 1000);
+    }
+  });
+
+  it('RESIZE_START on 10:07 entry → preview range snaps to 10:00-11:00', () => {
+    const { state } = dispatch(IDLE, {
+      type: 'RESIZE_START',
+      timeblockId: 'a',
+      direction: 'bottom',
+      point: origin,
+      originalPosition: off607,
+    });
+
+    if (state.mode === 'resizing') {
+      expect(state.previewTime.start.getHours()).toBe(10);
+      expect(state.previewTime.start.getMinutes()).toBe(0);
+      expect(state.previewTime.end.getHours()).toBe(11);
+      expect(state.previewTime.end.getMinutes()).toBe(0);
+    }
+  });
+
+  it('resize 10:07-11:07 entry を +15px → 10:00-11:15', () => {
+    const resizingState: InteractionState = {
+      mode: 'resizing',
+      timeblockId: 'a',
+      startPoint: origin,
+      currentPoint: origin,
+      originalPosition: off607,
+      direction: 'bottom',
+      snappedHeight: 60,
+      previewTime: {
+        start: new Date('2026-01-15T10:07:00'),
+        end: new Date('2026-01-15T11:07:00'),
+      },
+      isOverlapping: false,
+    };
+    const movedPoint = { clientX: origin.clientX, clientY: origin.clientY + 15 };
+    const { state } = dispatch(resizingState, { type: 'POINTER_MOVE', point: movedPoint });
+
+    if (state.mode === 'resizing') {
+      expect(state.snappedHeight).toBe(75);
+      expect(state.previewTime.start.getHours()).toBe(10);
+      expect(state.previewTime.start.getMinutes()).toBe(0);
+      expect(state.previewTime.end.getHours()).toBe(11);
+      expect(state.previewTime.end.getMinutes()).toBe(15);
+    }
+  });
+});
+
+// ========================================
+// Day-end boundary regression
+// ========================================
+
+describe('day-end boundary regression', () => {
+  it('dragging a 1-hour entry to 23:00 keeps the end at next-day 00:00', () => {
+    const draggingState: InteractionState = {
+      mode: 'dragging',
+      timeblockId: 'a',
+      startPoint: origin,
+      currentPoint: origin,
+      originalPosition: { top: 1320, left: 0, width: 200, height: 60 },
+      dateIndex: 0,
+      targetDateIndex: 0,
+      snappedTop: 1320,
+      previewTime: {
+        start: new Date('2026-01-15T22:00:00'),
+        end: new Date('2026-01-15T23:00:00'),
+      },
+      isOverlapping: false,
+    };
+    const movedPoint = { clientX: origin.clientX, clientY: origin.clientY + 60 };
+    const { state } = dispatch(draggingState, { type: 'POINTER_MOVE', point: movedPoint });
+
+    if (state.mode === 'dragging') {
+      expect(state.snappedTop).toBe(1380);
+      expect(state.previewTime.start.getHours()).toBe(23);
+      expect(state.previewTime.start.getMinutes()).toBe(0);
+      expect(state.previewTime.end.getDate()).toBe(16);
+      expect(state.previewTime.end.getHours()).toBe(0);
+      expect(state.previewTime.end.getMinutes()).toBe(0);
+      expect(state.previewTime.end.getTime() - state.previewTime.start.getTime()).toBe(
+        60 * 60 * 1000,
+      );
+    }
+  });
+
+  it('RESIZE_START supports a 23:00-24:00 range instead of collapsing to one interval', () => {
+    const { state } = dispatch(IDLE, {
+      type: 'RESIZE_START',
+      timeblockId: 'a',
+      direction: 'bottom',
+      point: origin,
+      originalPosition: { top: 1380, left: 0, width: 200, height: 60 },
+    });
+
+    if (state.mode === 'resizing') {
+      expect(state.snappedHeight).toBe(60);
+      expect(state.previewTime.start.getHours()).toBe(23);
+      expect(state.previewTime.end.getDate()).toBe(16);
+      expect(state.previewTime.end.getHours()).toBe(0);
+      expect(state.previewTime.end.getMinutes()).toBe(0);
+    }
+  });
+
+  it('grid selection at 23:45 ends at next-day 00:00', () => {
+    const { state } = dispatch(IDLE, {
+      type: 'GRID_POINTER_DOWN',
+      point: origin,
+      dateIndex: 0,
+      gridY: 1425,
+    });
+
+    if (state.mode === 'selecting') {
+      expect(state.selectionRange.start.getHours()).toBe(23);
+      expect(state.selectionRange.start.getMinutes()).toBe(45);
+      expect(state.selectionRange.end.getDate()).toBe(16);
+      expect(state.selectionRange.end.getHours()).toBe(0);
+      expect(state.selectionRange.end.getMinutes()).toBe(0);
+    }
+  });
+});
+
+// ========================================
+// Grid selection (GRID_POINTER_DOWN → POINTER_MOVE → POINTER_UP)
+// ========================================
