+++ apps/product/src/features/calendar/domain/interaction/machine.ts
@@ -3,9 +3,25 @@
  *
  * React/DOM依存ゼロ。すべての状態遷移とグリッド計算を純粋関数で実装。
  * テスト: expect(interactionReducer(state, action, ctx)).toEqual(...)
+ *
+ * 実装は責務ごとに分割している:
+ * - 定数: machine-constants.ts
+ * - グリッド幾何計算: grid-geometry.ts
+ * - POINTER_MOVE: pointer-move.ts
+ * - POINTER_UP: pointer-up.ts
  */
 
 import { DEFAULT_DRAG_SNAP_MINUTES } from '../precision';
+import {
+  buildDragTimeRange,
+  buildSelectionRange,
+  ensureEndAfterStartSnap,
+  resolveTargetDate,
+  snapEndToGrid,
+} from './grid-geometry';
+import { IDLE, LONGPRESS_DELAY_MS, SELECTION_LONGPRESS_DELAY_MS } from './machine-constants';
+import { handlePointerMove } from './pointer-move';
+import { handlePointerUp } from './pointer-up';
 import { snapToGrid } from './time-math';
 import type {
   InteractionAction,
@@ -13,138 +29,17 @@ import type {
   InteractionEffect,
   InteractionResult,
   InteractionState,
-  Point,
-  TimeRange,
 } from './types';
 
 // Re-export for consumers that import from machine.ts
+export {
+  DRAG_THRESHOLD_PX,
+  IDLE,
+  LONGPRESS_DELAY_MS,
+  SELECTION_LONGPRESS_DELAY_MS,
+  TOUCH_SCROLL_THRESHOLD_PX,
+} from './machine-constants';
 export { snapToGrid } from './time-math';
-
-// ========================================
-// Constants
-// ========================================
-
-/** マウスドラッグ起動閾値（px） */
-export const DRAG_THRESHOLD_PX = 5;
-
-/**
- * snappedTop を当日範囲内 [0, 24h - durationPx] に clamp する。
- * pixelsToTimeUnsnapped が時刻を 23:59 でクランプするため、ピクセルも合わせないと
- * ghost が画面外へずれる（Codex P2 指摘）。
- */
-function clampSnappedTopToDay(snappedTop: number, hourHeight: number, durationPx = 0): number {
-  const dayMaxPx = 24 * hourHeight - durationPx;
-  return Math.max(0, Math.min(snappedTop, Math.max(0, dayMaxPx)));
-}
-
-/** タッチ移動でロングプレスをキャンセルする閾値（px） — スクロール許容のため */
-export const TOUCH_SCROLL_THRESHOLD_PX = 10;
-
-/** イベントドラッグのロングプレス遅延（ms） */
-export const LONGPRESS_DELAY_MS = 500;
-
-/** グリッド選択のロングプレス遅延（ms） */
-export const SELECTION_LONGPRESS_DELAY_MS = 300;
-
-/** アイドル状態の初期値 */
-export const IDLE: InteractionState = { mode: 'idle' };
-
-// ========================================
-// Helpers
-// ========================================
-
-function maxAbsDelta(a: Point, b: Point): number {
-  return Math.max(Math.abs(a.clientX - b.clientX), Math.abs(a.clientY - b.clientY));
-}
-
-type GridSnap = ReturnType<typeof snapToGrid>;
-
-function snapEndToGrid(yPx: number, hourHeight: number, intervalMin: number): GridSnap {
-  const pxPerInterval = (hourHeight / 60) * intervalMin;
-  if (pxPerInterval <= 0) return snapToGrid(yPx, hourHeight, intervalMin);
-
-  const dayHeight = 24 * hourHeight;
-  const clampedY = Math.max(0, Math.min(yPx, dayHeight));
-  const snappedTop = Math.max(
-    0,
-    Math.min(dayHeight, Math.round(clampedY / pxPerInterval) * pxPerInterval),
-  );
-  const totalMinutes = Math.min(24 * 60, Math.round((snappedTop / hourHeight) * 60));
-
-  return {
-    snappedTop,
-    hour: Math.floor(totalMinutes / 60),
-    minute: totalMinutes % 60,
-  };
-}
-
-function ensureEndAfterStartSnap(
-  startSnap: GridSnap,
-  endSnap: GridSnap,
-  hourHeight: number,
-  intervalMin: number,
-): GridSnap {
-  if (endSnap.snappedTop > startSnap.snappedTop) return endSnap;
-
-  const minEndTop = Math.min(
-    24 * hourHeight,
-    startSnap.snappedTop + (hourHeight / 60) * intervalMin,
-  );
-  return snapEndToGrid(minEndTop, hourHeight, intervalMin);
-}
-
-/** Dragging snaps both boundaries so drop time and visual length stay on the active grid. */
-function buildDragTimeRange(
-  targetDate: Date,
-  startSnap: GridSnap,
-  endSnap: GridSnap,
-  intervalMin: number,
-): TimeRange {
-  const start = new Date(targetDate);
-  start.setHours(startSnap.hour, startSnap.minute, 0, 0);
-  const end = new Date(targetDate);
-  end.setHours(endSnap.hour, endSnap.minute, 0, 0);
-
-  if (end.getTime() <= start.getTime()) {
-    end.setTime(start.getTime() + intervalMin * 60_000);
-  }
-
-  return { start, end };
-}
-
-/** Resolve the target date for a given date index */
-function resolveTargetDate(ctx: InteractionContext, targetDateIndex: number): Date {
-  if (ctx.viewMode !== 'day' && ctx.displayDates?.[targetDateIndex]) {
-    return ctx.displayDates[targetDateIndex];
-  }
-  return ctx.date;
-}
-
-/** Build a time range for a grid selection (downward only from startY) */
-function buildSelectionRange(
-  startY: number,
-  endY: number,
-  hourHeight: number,
-  targetDate: Date,
-  intervalMin: number,
-): TimeRange {
-  const startSnap = snapToGrid(startY, hourHeight, intervalMin);
-  // 下方向のみ: endY が startY より上なら startY に固定
-  const clampedEndY = Math.max(endY, startY);
-  let endSnap = snapEndToGrid(clampedEndY, hourHeight, intervalMin);
-  endSnap = ensureEndAfterStartSnap(startSnap, endSnap, hourHeight, intervalMin);
-
-  const start = new Date(targetDate);
-  start.setHours(startSnap.hour, startSnap.minute, 0, 0);
-  const end = new Date(targetDate);
-  end.setHours(endSnap.hour, endSnap.minute, 0, 0);
-
-  return { start, end };
-}
-
-// ========================================
-// Reducer
-// ========================================
 
 /**
  * インタラクション状態機械の純粋レデューサー
@@ -374,271 +269,5 @@ export function interactionReducer(
 
     default:
       return { state, effects };
-  }
-}
-
-// ========================================
-// POINTER_MOVE handler (per-mode routing)
-// ========================================
-
-function handlePointerMove(
-  state: InteractionState,
-  action: { type: 'POINTER_MOVE'; point: Point; targetDateIndex?: number },
-  ctx: InteractionContext,
-  effects: InteractionEffect[],
-  interval: number,
-): InteractionResult {
-  switch (state.mode) {
-    case 'pending': {
-      if (maxAbsDelta(state.startPoint, action.point) <= DRAG_THRESHOLD_PX) {
-        return { state, effects };
-      }
-      // Threshold crossed → transition to dragging
-      const deltaY = action.point.clientY - state.startPoint.clientY;
-      const durationMs = ctx.getTimeblockDurationMs(state.timeblockId);
-      const durationPx = (durationMs / 60_000) * (ctx.hourHeight / 60);
-      const rawTop = clampSnappedTopToDay(
-        state.originalPosition.top + deltaY,
-        ctx.hourHeight,
-        durationPx,
-      );
-      const startSnap = snapToGrid(rawTop, ctx.hourHeight, interval);
-      const endSnap = snapEndToGrid(rawTop + durationPx, ctx.hourHeight, interval);
-      const targetDateIndex = action.targetDateIndex ?? state.dateIndex;
-      const targetDate = resolveTargetDate(ctx, targetDateIndex);
-      const previewTime = buildDragTimeRange(targetDate, startSnap, endSnap, interval);
-      const isOverlapping = ctx.checkOverlap(
-        state.timeblockId,
-        previewTime.start,
-        previewTime.end,
-        'drag',
-      );
-
-      effects.push({
-        type: 'DRAG_STORE_START',
-        timeblockId: state.timeblockId,
-        dateIndex: state.dateIndex,
-      });
-
-      return {
-        state: {
-          mode: 'dragging',
-          timeblockId: state.timeblockId,
-          startPoint: state.startPoint,
-          currentPoint: action.point,
-          originalPosition: state.originalPosition,
-          dateIndex: state.dateIndex,
-          targetDateIndex,
-          snappedTop: startSnap.snappedTop,
-          previewTime,
-          isOverlapping,
-        },
-        effects,
-      };
-    }
-
-    case 'longpress-pending': {
-      if (maxAbsDelta(state.startPoint, action.point) > TOUCH_SCROLL_THRESHOLD_PX) {
-        effects.push({ type: 'CLEAR_LONGPRESS_TIMER' });
-        return { state: IDLE, effects };
-      }
-      return { state, effects };
-    }
-
-    case 'dragging': {
-      const deltaY = action.point.clientY - state.startPoint.clientY;
-      const durationMs = ctx.getTimeblockDurationMs(state.timeblockId);
-      const durationPx = (durationMs / 60_000) * (ctx.hourHeight / 60);
-      const rawTop = clampSnappedTopToDay(
-        state.originalPosition.top + deltaY,
-        ctx.hourHeight,
-        durationPx,
-      );
-      const startSnap = snapToGrid(rawTop, ctx.hourHeight, interval);
-      const endSnap = snapEndToGrid(rawTop + durationPx, ctx.hourHeight, interval);
-      const targetDateIndex = action.targetDateIndex ?? state.targetDateIndex;
-      const targetDate = resolveTargetDate(ctx, targetDateIndex);
-      const previewTime = buildDragTimeRange(targetDate, startSnap, endSnap, interval);
-      const isOverlapping = ctx.checkOverlap(
-        state.timeblockId,
-        previewTime.start,
-        previewTime.end,
-        'drag',
-      );
-
-      if (startSnap.snappedTop !== state.snappedTop) {
-        effects.push({ type: 'HAPTIC', pattern: 'tap' });
-      }
-      effects.push({ type: 'DRAG_STORE_UPDATE', targetDateIndex });
-
-      return {
-        state: {
-          ...state,
-          currentPoint: action.point,
-          targetDateIndex,
-          snappedTop: startSnap.snappedTop,
-          previewTime,
-          isOverlapping,
-        },
-        effects,
-      };
-    }
-
-    case 'resizing': {
-      const deltaY = action.point.clientY - state.startPoint.clientY;
-      const minHeight = (ctx.hourHeight / 60) * interval;
-      const resizeMinEndMinutes = ctx.getResizeMinEndMinutes?.(state.timeblockId) ?? null;
-      const resizeMinEndTop =
-        resizeMinEndMinutes == null
-          ? 0
-          : (Math.ceil(resizeMinEndMinutes / interval) * interval * ctx.hourHeight) / 60;
-      // upper cap: end が当日内に収まる範囲
-      const startSnap = snapToGrid(state.originalPosition.top, ctx.hourHeight, interval);
-      const maxHeight = Math.max(minHeight, 24 * ctx.hourHeight - startSnap.snappedTop);
-      const rawEndTop = Math.min(
-        24 * ctx.hourHeight,
-        Math.max(
-          startSnap.snappedTop + minHeight,
-          resizeMinEndTop,
-          state.originalPosition.top + state.originalPosition.height + deltaY,
-        ),
-      );
-      const endSnap = snapEndToGrid(rawEndTop, ctx.hourHeight, interval);
-      const newHeight = Math.min(
-        maxHeight,
-        Math.max(minHeight, endSnap.snappedTop - startSnap.snappedTop),
-      );
-
-      if (newHeight !== state.snappedHeight) {
-        effects.push({ type: 'HAPTIC', pattern: 'tap' });
-      }
-
-      const start = new Date(ctx.date);
-      start.setHours(startSnap.hour, startSnap.minute, 0, 0);
-      const end = new Date(ctx.date);
-      end.setHours(endSnap.hour, endSnap.minute, 0, 0);
-
-      const previewTime: TimeRange = { start, end };
-      const isOverlapping = ctx.checkOverlap(state.timeblockId, start, end, 'resize');
-
-      return {
-        state: {
-          ...state,
-          currentPoint: action.point,
-          snappedHeight: newHeight,
-          previewTime,
-          isOverlapping,
-        },
-        effects,
-      };
-    }
-
-    case 'selecting': {
-      const deltaY = action.point.clientY - state.startPoint.clientY;
-      const currentGridY = state.gridStartY + deltaY;
-      const targetDate = resolveTargetDate(ctx, state.dateIndex);
-      const selectionRange = buildSelectionRange(
-        state.gridStartY,
-        currentGridY,
-        ctx.hourHeight,
-        targetDate,
-        interval,
-      );
-
-      return {
-        state: {
-          ...state,
-          currentPoint: action.point,
-          selectionRange,
-        },
-        effects,
-      };
-    }
-
-    case 'selection-longpress-pending': {
-      if (maxAbsDelta(state.startPoint, action.point) > TOUCH_SCROLL_THRESHOLD_PX) {
-        effects.push({ type: 'CLEAR_LONGPRESS_TIMER' });
-        return { state: IDLE, effects };
-      }
-      return { state, effects };
-    }
-
-    default:
-      return { state, effects };
-  }
-}
-
-// ========================================
-// POINTER_UP handler (per-mode completion)
-// ========================================
-
-function handlePointerUp(state: InteractionState, effects: InteractionEffect[]): InteractionResult {
-  switch (state.mode) {
-    case 'pending': {
-      effects.push({ type: 'EVENT_CLICK', timeblockId: state.timeblockId });
-      return { state: IDLE, effects };
-    }
-
-    case 'longpress-pending': {
-      effects.push({ type: 'CLEAR_LONGPRESS_TIMER' });
-      effects.push({ type: 'EVENT_CLICK', timeblockId: state.timeblockId });
-      return { state: IDLE, effects };
-    }
-
-    case 'dragging': {
-      effects.push({ type: 'DRAG_STORE_END' });
-
-      if (state.isOverlapping) {
-        effects.push({ type: 'DROP_REJECTED', timeblockId: state.timeblockId, reason: 'overlap' });
-        effects.push({ type: 'HAPTIC', pattern: 'error' });
-      } else {
-        effects.push({
-          type: 'DROP',
-          timeblockId: state.timeblockId,
-          time: state.previewTime,
-          targetDateIndex: state.targetDateIndex,
-        });
-      }
-
-      return { state: IDLE, effects };
-    }
-
-    case 'resizing': {
-      if (state.isOverlapping) {
-        effects.push({
-          type: 'RESIZE_REJECTED',
-          timeblockId: state.timeblockId,
-          reason: 'overlap',
-        });
-        effects.push({ type: 'HAPTIC', pattern: 'error' });
-      } else {
-        effects.push({
-          type: 'RESIZE_COMPLETE',
-          timeblockId: state.timeblockId,
-          time: state.previewTime,
-        });
-      }
-      return { state: IDLE, effects };
-    }
-
-    case 'selecting': {
-      const hasMoved = maxAbsDelta(state.startPoint, state.currentPoint) > DRAG_THRESHOLD_PX;
-      if (hasMoved && !state.isOverlapping) {
-        effects.push({
-          type: 'SELECT_COMPLETE',
-          dateIndex: state.dateIndex,
-          range: state.selectionRange,
-        });
-      }
-      return { state: IDLE, effects };
-    }
-
-    case 'selection-longpress-pending': {
-      effects.push({ type: 'CLEAR_LONGPRESS_TIMER' });
-      return { state: IDLE, effects };
-    }
-
-    default:
-      return { state: IDLE, effects };
   }
 }
